requirements.txt
azure-common==1.1.9
azure-mgmt-compute==4.0.1
azure-mgmt-network==2.0.1
azure-mgmt==4.0.0
azure-nspkg
boto==2.34.0
commandr==1.3.2
flask==1.0.2
Flask-RESTful==0.3.5
gevent==1.0.2
Jinja2==2.10
kazoo==2.0
librato-metrics==0.7.0
psutil==2.2.1
python-dateutil==2.4.2
raven==5.1.1
requests==2.16
